Themis: add support for allocation repeats with Flux
Problem: the Flux ResourceManager class needs to determine whether the
Flux instance it is communicating with is a system instance or just a
single-user instance launched under Slurm or LSF. If the latter, requests
for allocations need to be routed to Slurm/LSF, since Flux doesn't own any
resources. The Themis backend needs to be able to make the same distinction
when requesting allocation repeats, but the logic was not implemented.

Solution: implement logic so that the Themis backend can distinguish between
Flux system instances and single-user instances.